Added .gcode to gitignore, changed infill slightly
Infill previously had a check to remove duplicate points. As far as I
can tell, duplicate points only occur when the border of the domain
intersects an infill line at a vertex of the domain. In this scenario,
both points are valid because otherwise the "no odd numbers of
intersections" rule will throw that infill line away.