Removed weird array work in Surface3D.overlap
I haven't profiled this but it's weird to use an array and resize it
myself in one place and arraylists everywhere else.

This change was precipitated by an attempt to use interval trees for
slicing, but that turns out to be really complicated with nonplanar
surfaces because calculating the minimum and maximum vertical distance
between a triangle and a surface is expensive.