fixed dll, rearranged java src
The string interface for Clipper was missing a space after the : polygon
delimiter. In the previous commit I fixed this for the Domain java
struct, this commit fixes the same issue in the C++ string exporter.

I also rearranged the java source to make src\process less of an
"everything" folder.