Added Point6D, modified reproject to carry normal vectors through
Modified Extrusion3D to store a normal vector for its start point
and end point. Modified the projection code to use a new primitive
Point6D which subclasses Point3D such that the normal vector of the
triangle a point projects onto is stored in the resulting Point6D.
Normal vector defaults to zero in the standard constructor.